Refactor ShadowStylePickerView
